+++ include/globals.h
@@ -42,3 +42,6 @@ extern byte preset;
 
 extern byte arp_output_note;
 extern optional<byte> control_voltage_note;
+
+// voice_index[operator] tells you the preset's voice to read the operator's settings from.
+extern byte* voice_index; // array of size 6, set depending on preset.paraphonic