+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 2.1] - TODO
+
+### Added
+- 2 and 6 voice polyphonic modes (via F5 fat mode).
+- 3-voice 2-operator polyphonic mode (via F6 fat mode in paraphonic mode).
+
+### Fixed
+
+### Changed
+
+### Removed
+
 ## [Unreleased 2.0] - 2023-02-14
 
 ### Fixed